unified server file + discollection handling
all the code for the server side is now unified inside a single file. Functions to handle clients disconnections from the server have been added. 
The new functions work correctly except from an error that occurs if the clients get kicked out and then someone reconnects: this raises an error every time the server receives a message and every time the disconnection and reconnection occurs, the server seems to receive 1 additional copy of the message sent.
The app still functions but adds a copy of the messages received every time a client gets kicked out.
Mqtt communications is still not handled properly